+++ src/components/snack/SnackList.module.css
@@ -1,0 +1,98 @@
+/* 전체 그리드 컨테이너 */
+.gridContainer {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr); /* 기본 4열 */
+  gap: 20px;
+  width: 100%;
+}
+
+/* 개별 카드 */
+.card {
+  border: 1px solid #eee;
+  border-radius: 12px;
+  padding: 20px;
+  background-color: #fff;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  transition: transform 0.2s, box-shadow 0.2s; /* 부드러운 움직임 추가 */
+}
+
+/* 카드에 마우스 올렸을 때 살짝 떠오르는 효과 */
+.card:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+/* 이미지 영역 (회색 박스) */
+.imagePlaceholder {
+  width: 100%;
+  height: 150px;
+  background-color: #f0f0f0;
+  border-radius: 8px;
+  margin-bottom: 15px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #aaa;
+  font-size: 12px;
+}
+
+/* 텍스트 영역 */
+.snackName {
+  font-size: 18px;
+  font-weight: bold;
+  margin: 0 0 10px 0;
+  color: #333;
+}
+
+.infoRow {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 15px;
+}
+
+.price {
+  font-size: 20px;
+  font-weight: bold;
+  color: #ff9f1c; /* 포인트 컬러 */
+}
+
+.stock {
+  font-size: 12px;
+  color: #888;
+}
+
+/* 주문 버튼 */
+.orderButton {
+  width: 100%;
+  padding: 12px;
+  background-color: #ff9f1c;
+  color: white;
+  border: none;
+  border-radius: 8px;
+  font-weight: bold;
+  font-size: 16px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+.orderButton:hover {
+  background-color: #e68a00; /* 호버 시 조금 더 진한 색 */
+}
+
+/* 반응형: 화면 폭이 768px 이하(태블릿/모바일)면 2열로 변경 */
+@media (max-width: 768px) {
+  .gridContainer {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+/* 반응형: 화면 폭이 480px 이하(작은 모바일)면 1열로 변경 */
+@media (max-width: 480px) {
+  .gridContainer {
+    grid-template-columns: 1fr;
+  }
+}
